port to QGIS3 (#14)
* port to QGIS3

initial port to QGIS3.  PDF and print don't seem to work, and it could be cleaned up a little.

* Fix export to pdf for QGIS3

* Fix printing for QGIS3

But why is the print rasterised?  It wasn't in QGIS2...

* Fix usage of QgsMessageBar

* Bugfix - don't show button to open output file if the operation was cancelled

I guess we need to check if there are any other places it is shown when it shouldn't be.

* Don't claim to be for QGIS 2+

* Consistently use "from qgis.PyQt"

This might be helpful one day...

* Edit resource_rc.py even though we're not supposed to

* Update readme references to old QGIS versions